corrected auth 3

+++ app/Http/Controllers/TasksController.php
@@ -75,10 +75,12 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::findOrFail($id);
-        
+        if (\Auth::id() === $task->user_id) {
         return view('tasks.show', [
             'task' => $task,
-    ]);
+            ]);
+        }
+        return redirect ('/');
     }
 
     /**
@@ -90,11 +92,13 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::findOrFail($id);
-       
+        if (\Auth::id() === $task->user_id) {
         return view('tasks.edit', [
             'task' => $task,
             ]);
         }
+        return redirect ('/');
+    }
 
     /**
      * Update the specified resource in storage.
@@ -105,16 +109,16 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
-         $request->validate([
+        $request->validate([
             'status' => 'required|max:10',
             'content' => 'required|max:255',
             ]);
-         $task = Task::findOrFail($id);
-         $task->status = $request->status;
-         $task->content = $request->content;
-         $task->save();
+        $task = Task::findOrFail($id);
+        $task->status = $request->status;
+        $task->content = $request->content;
+        $task->save();
          
-         return redirect('/');
+        return redirect('/');
     }
 
     /**
@@ -125,9 +129,10 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-         $task = Task::findOrFail($id);
-         
-         $task->delete();
-        
+        $task = Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
+        $task->delete();
+        }
+        return redirect ('/');
     }
 }
